backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "google-generativeai>=0.8.6",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.41.0",
    "pymongo[srv]>=4.12.1",
    "bcrypt>=4.3.0",
    "pyjwt>=2.10.1",
    "certifi>=2026.2.25",
    "httpx>=0.28.1",
]
